requirements.txt
# Requirements for this package
pyodbc
python-dotenv
sqlalchemy>2.0
